Script to render segmentation masks
The original repository contains a script to render the segmentation
masks of all YCB objects in the scene plus the hand on top of the RGB
color image. Instead, we provide a script to generate a single channel
segmentation mask of only the hand and interacting object in the scene.